# ios/Sources/LaughTrackAPIClient/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LaughTrack API",
    "version": "1.0.0",
    "description": "LaughTrack v1 REST API — comedy shows, clubs, comedians, and favorites."
  },
  "servers": [
    {
      "url": "https://www.laugh-track.com/api/v1",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/home/feed": {
      "get": {
        "operationId": "getHomeFeed",
        "summary": "Composite home-screen feed with curated and personalized sections",
        "description": "Returns hero context (zip/city/state + up to 3 near-you shows), curated discovery arrays, optional podcast episode recommendations, and followedComedianShows for authenticated profiles. Signed-out callers receive anonymous episode discovery and an empty followedComedianShows array. Rate limit: 60 req/min anon, 300 req/min authenticated. Cache-Control: private, max-age=60 — response is personalized by profile, session zipCode, and Vercel geo-IP, so shared CDN caching is disabled.",
        "tags": ["home"],
        "security": [{ "bearerAuth": [] }, {}],
        "parameters": [
          {
            "name": "zip",
            "in": "query",
            "required": false,
            "description": "Optional 5-digit US zip override. When supplied, beats the signed-in user's profile zipCode for this request — used for anonymous callers or profile-preview.",
            "schema": { "type": "string", "pattern": "^\\d{5}$" }
          },
          {
            "name": "distance",
            "in": "query",
            "required": false,
            "description": "Radius in miles for zip-scoped recommendations (1-500, default 25).",
            "schema": { "type": "integer", "minimum": 1, "maximum": 500, "default": 25 }
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "description": "Client platform used to select the server-directed Discover rail policy. Defaults to web for backward compatibility.",
            "schema": {
              "type": "string",
              "enum": ["web", "ios", "android"],
              "default": "web"
            }
          },
          {
            "name": "X-Timezone",
            "in": "header",
            "description": "IANA timezone identifier (defaults to UTC)",
            "schema": { "type": "string", "default": "UTC" }
          }
        ],
        "responses": {
          "200": {
            "description": "Home feed payload",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HomeFeedResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid zip parameter",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived (15-minute) access JWT obtained from POST /auth/token or /auth/refresh"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "SocialData": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": { "type": "integer" },
          "instagramAccount": { "type": ["string", "null"] },
          "instagramFollowers": { "type": ["integer", "null"] },
          "tiktokAccount": { "type": ["string", "null"] },
          "tiktokFollowers": { "type": ["integer", "null"] },
          "youtubeAccount": { "type": ["string", "null"] },
          "youtubeFollowers": { "type": ["integer", "null"] },
          "website": { "type": ["string", "null"] },
          "popularity": { "type": ["number", "null"] },
          "linktree": { "type": ["string", "null"] }
        }
      },
...
      "Ticket": {
        "type": "object",
        "properties": {
          "price": { "type": ["number", "null"] },
          "purchaseUrl": { "type": ["string", "null"] },
          "soldOut": { "type": ["boolean", "null"] },
          "type": { "type": ["string", "null"] }
        }
      },
      "ComedianLineup": {
        "type": "object",
        "required": ["name", "imageUrl", "uuid", "id"],
        "properties": {
          "name": { "type": "string" },
          "imageUrl": { "type": "string" },
          "uuid": { "type": "string" },
          "id": { "type": "integer" },
          "userId": { "type": "integer" },
          "socialData": { "$ref": "#/components/schemas/SocialData" },
          "isFavorite": { "type": "boolean" },
          "showCount": { "type": "integer" },
          "role": { "type": ["string", "null"] },
          "isAlias": { "type": "boolean" },
          "parentComedian": { "$ref": "#/components/schemas/ComedianLineup" },
          "lineupItems": {
            "type": "array",
            "items": {}
          }
        }
      },
      "Tag": {
        "type": "object",
        "required": ["slug", "name"],
        "properties": {
          "slug": { "type": "string" },
          "name": { "type": "string" }
        }
      },
      "Show": {
        "type": "object",
        "required": ["id", "clubId", "date", "imageUrl"],
        "properties": {
          "id": { "type": "integer" },
          "clubId": { "type": "integer" },
          "clubName": { "type": "string" },
          "clubCity": { "type": ["string", "null"] },
          "clubState": { "type": ["string", "null"] },
          "date": { "type": "string", "format": "date-time" },
          "tickets": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Ticket" }
          },
          "name": { "type": ["string", "null"] },
          "popularityScore": { "type": ["number", "null"] },
          "socialData": { "$ref": "#/components/schemas/SocialData" },
          "lineup": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ComedianLineup" }
          },
          "tags": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Tag" }
          },
          "description": { "type": "string" },
          "address": { "type": "string" },
          "room": { "type": ["string", "null"] },
          "imageUrl": { "type": "string" },
          "soldOut": { "type": "boolean" },
          "distanceMiles": { "type": ["number", "null"] },
          "timezone": { "type": ["string", "null"] }
        }
      },
...
      "ClubListItem": {
        "type": "object",
        "required": ["id", "address", "name", "imageUrl", "activeComedianCount"],
        "properties": {
          "id": { "type": "integer" },
          "address": { "type": "string" },
          "name": { "type": "string" },
          "zipCode": { "type": ["string", "null"] },
          "imageUrl": { "type": "string" },
          "activeComedianCount": { "type": "integer" }
        }
      },
...
      "ComedianListItem": {
        "type": "object",
        "required": ["id", "uuid", "name", "imageUrl", "socialData", "showCount"],
        "properties": {
          "id": { "type": "integer" },
          "uuid": { "type": "string" },
          "name": { "type": "string" },
          "imageUrl": { "type": "string" },
          "socialData": { "$ref": "#/components/schemas/SocialData" },
          "showCount": { "type": "integer" }
        }
      },
...
      "HomeFeedHero": {
        "type": "object",
        "required": ["shows"],
        "properties": {
          "zipCode": { "type": ["string", "null"] },
          "city": { "type": ["string", "null"] },
          "state": { "type": ["string", "null"] },
          "shows": {
            "type": "array",
            "description": "Up to 3 shows near the hero zip, for the top-of-home carousel.",
            "items": { "$ref": "#/components/schemas/Show" }
          }
        }
      },
      "HomeFeedPodcast": {
        "type": "object",
        "required": ["id", "slug", "title", "episodeCount"],
        "properties": {
          "id": { "type": "integer" },
          "slug": { "type": "string" },
          "title": { "type": "string" },
          "authorName": { "type": ["string", "null"] },
          "websiteUrl": { "type": ["string", "null"] },
          "feedUrl": { "type": ["string", "null"] },
          "imageUrl": { "type": ["string", "null"] },
          "description": { "type": ["string", "null"] },
          "episodeCount": { "type": "integer" }
        }
      },
      "HomeFeedPodcastEpisodePodcast": {
        "type": "object",
        "required": ["id", "slug", "title", "imageUrl"],
        "properties": {
          "id": { "type": "integer" },
          "slug": { "type": "string" },
          "title": { "type": "string" },
          "imageUrl": { "type": ["string", "null"] }
        }
      },
      "HomeFeedPodcastEpisodeComedian": {
        "type": "object",
        "required": ["id", "uuid", "name", "imageUrl"],
        "properties": {
          "id": { "type": "integer" },
          "uuid": { "type": "string" },
          "name": { "type": "string" },
          "imageUrl": { "type": "string" }
        }
      },
      "HomeFeedPodcastEpisodeRecommendation": {
        "type": "object",
        "required": [
          "reason",
          "comedian",
          "appearanceRole",
          "followedComedian",
          "favoritePodcast"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "enum": [
              "followed_comedian",
              "favorite_podcast",
              "guest_appearance",
              "popular_comedian",
              "recent_episode"
            ]
          },
          "comedian": {
            "$ref": "#/components/schemas/HomeFeedPodcastEpisodeComedian"
          },
          "appearanceRole": {
            "type": "string",
            "enum": ["host", "cohost", "guest"]
          },
          "followedComedian": { "type": "boolean" },
          "favoritePodcast": { "type": "boolean" }
        }
      },
      "HomeFeedPodcastEpisode": {
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "releaseDate",
          "durationSeconds",
          "episodeUrl",
          "audioUrl",
          "podcast",
          "recommendation"
        ],
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "releaseDate": { "type": "string", "format": "date-time" },
          "durationSeconds": { "type": ["integer", "null"] },
          "episodeUrl": { "type": ["string", "null"] },
          "audioUrl": { "type": ["string", "null"] },
          "podcast": {
            "$ref": "#/components/schemas/HomeFeedPodcastEpisodePodcast"
          },
          "recommendation": {
            "$ref": "#/components/schemas/HomeFeedPodcastEpisodeRecommendation"
          }
        }
      },
      "HomeFeedDynamicRailPerformer": {
        "type": "object",
        "required": ["id", "uuid", "name"],
        "properties": {
          "id": { "type": "integer" },
          "uuid": { "type": "string" },
          "name": { "type": "string" }
        }
      },
      "HomeFeedDynamicRailPodcast": {
        "type": "object",
        "required": ["id", "slug", "title"],
        "properties": {
          "id": { "type": "integer" },
          "slug": { "type": "string" },
          "title": { "type": "string" }
        }
      },
      "HomeFeedDynamicRailFreshnessProvenance": {
        "type": "object",
        "required": ["kind"],
        "properties": {
          "kind": {
            "type": "string",
            "enum": ["laughtrack_observation", "verified_announcement"]
          },
          "source": { "type": "string" },
          "announcedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "HomeFeedDynamicRailFavoriteSignals": {
        "type": "object",
        "required": ["recentCount", "baselineCount", "growth", "confidence"],
        "properties": {
          "recentCount": { "type": "integer" },
          "baselineCount": { "type": "integer" },
          "growth": { "type": ["number", "null"] },
          "confidence": { "type": "number" }
        }
      },
...
      "HomeFeedDynamicRailReasonEvidence": {
        "type": "object",
        "description": "Typed evidence superset. Fields are populated according to the reason kind.",
        "properties": {
          "canonicalComedianId": { "type": "integer" },
          "localAppearanceCount": { "type": "integer" },
          "horizonDays": { "type": "integer" },
          "runStart": { "type": "string", "format": "date-time" },
          "runEnd": { "type": "string", "format": "date-time" },
          "homeMarket": {
            "type": ["object", "null"],
            "required": ["city", "state"],
            "properties": {
              "city": { "type": "string" },
              "state": { "type": "string" },
              "country": { "type": ["string", "null"] }
            }
          },
          "requestedMarket": {
            "type": ["object", "null"],
            "required": ["city", "state"],
            "properties": {
              "city": { "type": "string" },
              "state": { "type": "string" },
              "country": { "type": ["string", "null"] }
            }
          },
          "priorLocalAppearanceCount": { "type": "integer" },
          "lastLocalAppearanceAt": { "type": ["string", "null"], "format": "date-time" },
          "historyCoverageStart": { "type": ["string", "null"], "format": "date-time" },
          "historyCoverageShowCount": { "type": "integer" },
          "firstDiscoveredAt": { "type": ["string", "null"], "format": "date-time" },
          "freshnessProvenance": { "$ref": "#/components/schemas/HomeFeedDynamicRailFreshnessProvenance" },
          "featureVersion": { "type": ["string", "null"] },
          "featureAsOf": { "type": ["string", "null"], "format": "date-time" },
          "prominence": { "type": ["number", "null"] },
          "momentum": { "type": ["number", "null"] },
          "growth": { "type": ["number", "null"] },
          "confidence": { "type": ["number", "null"] },
          "availability": {
            "type": ["string", "null"],
            "enum": ["available", "unknown", "unavailable", null]
          },
          "signals": {
            "type": ["object", "null"],
            "required": ["behavior", "favorites", "social", "confidenceReasons"],
            "properties": {
              "behavior": { "$ref": "#/components/schemas/HomeFeedDynamicRailBehaviorSignals" },
              "favorites": { "$ref": "#/components/schemas/HomeFeedDynamicRailFavoriteSignals" },
              "social": { "$ref": "#/components/schemas/HomeFeedDynamicRailSocialSignals" },
              "confidenceReasons": {
                "type": "array",
                "items": { "type": "string" }
              }
            }
          },
          "podcast": { "$ref": "#/components/schemas/HomeFeedDynamicRailPodcast" },
          "attribution": {
            "type": "string",
            "enum": ["host", "cohost", "recent_appearance"]
          },
          "appearanceRole": {
            "type": ["string", "null"],
            "enum": ["host", "cohost", "guest", null]
          },
          "episode": {
            "type": ["object", "null"],
            "required": ["id", "title", "releaseDate"],
            "properties": {
              "id": { "type": "integer" },
              "title": { "type": "string" },
              "releaseDate": { "type": "string", "format": "date-time" }
            }
          }
        }
      },
      "HomeFeedDynamicRailReason": {
        "type": "object",
        "required": ["kind", "label", "evidence"],
        "properties": {
          "kind": {
            "type": "string",
            "description": "Extensible reason identifier. Clients must tolerate values introduced by newer recommendation policies."
          },
          "label": { "type": "string" },
          "evidence": { "$ref": "#/components/schemas/HomeFeedDynamicRailReasonEvidence" }
        }
      },
      "HomeFeedDynamicRailItem": {
        "type": "object",
        "required": ["id", "show", "reason"],
        "properties": {
          "id": { "type": "integer" },
          "show": { "$ref": "#/components/schemas/Show" },
          "performer": { "$ref": "#/components/schemas/HomeFeedDynamicRailPerformer" },
          "podcast": { "$ref": "#/components/schemas/HomeFeedDynamicRailPodcast" },
          "reason": { "$ref": "#/components/schemas/HomeFeedDynamicRailReason" }
        }
      },
      "HomeFeedDynamicRail": {
        "type": "object",
        "required": ["railKey", "label", "items"],
        "properties": {
          "railKey": {
            "type": "string",
            "description": "Extensible dynamic rail identifier. Clients must tolerate values introduced by newer recommendation policies."
          },
          "label": { "type": "string" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/HomeFeedDynamicRailItem" }
          }
        }
      },
      "HomeFeed": {
        "type": "object",
        "required": [
          "hero",
          "trendingComedians",
          "comediansNearYou",
          "showsTonight",
          "moreNearYou",
          "trendingThisWeek",
          "followedComedianShows",
          "trendingPodcasts",
          "popularClubs"
        ],
        "properties": {
          "hero": { "$ref": "#/components/schemas/HomeFeedHero" },
          "trendingComedians": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ComedianListItem" }
          },
          "comediansNearYou": {
            "type": "array",
            "description": "Empty when no zipCode is resolved (no profile zip, no ?zip param, no geo-IP).",
            "items": { "$ref": "#/components/schemas/ComedianListItem" }
          },
          "showsTonight": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Show" }
          },
          "moreNearYou": {
            "type": "array",
            "description": "Shows near the hero zip beyond the first 3 surfaced in hero.shows.",
            "items": { "$ref": "#/components/schemas/Show" }
          },
          "trendingThisWeek": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Show" }
          },
          "followedComedianShows": {
            "type": "array",
            "description": "Up to 8 upcoming shows featuring comedians followed by the authenticated profile, excluding shows already represented in higher-priority home-feed sections. Empty for signed-out callers.",
            "items": { "$ref": "#/components/schemas/Show" }
          },
          "podcastEpisodes": {
            "type": "array",
            "description": "Optional fresh episode recommendations ranked from comedian follows, podcast favorites, appearance role, comedian popularity, and recency. Older clients and stale cached payloads may omit this field.",
            "items": {
              "$ref": "#/components/schemas/HomeFeedPodcastEpisode"
            }
          },
          "trendingPodcasts": {
            "type": "array",
            "description": "Podcasts featuring local comedians when a hero zip is resolved, otherwise globally popular comedian-owned podcasts.",
            "items": { "$ref": "#/components/schemas/HomeFeedPodcast" }
          },
          "popularClubs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ClubListItem" }
          },
          "dynamicRails": {
            "type": "array",
            "description": "Non-empty dynamic show rail payloads with structured recommendation reasons. Use railPlan.itemIds for the policy-selected subset and order.",
            "items": { "$ref": "#/components/schemas/HomeFeedDynamicRail" }
          },
          "railPlan": {
            "$ref": "#/components/schemas/HomeFeedRailPlan"
          }
        }
      },
      "HomeFeedRailPlanEntry": {
        "type": "object",
        "required": ["railKey", "payloadKey", "position", "itemIds"],
        "properties": {
          "railKey": {
            "type": "string",
            "description": "Extensible rail identifier selected by the server policy."
          },
          "payloadKey": {
            "type": "string",
            "description": "Extensible HomeFeed payload field containing the rail data."
          },
          "position": { "type": "integer", "minimum": 0 },
          "itemIds": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "HomeFeedRailPlan": {
        "type": "object",
        "required": ["version", "catalogVersion", "policyVersion", "platform", "cycleIndex", "rails"],
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1,
            "description": "Extensible rail-plan schema version. Clients must tolerate newer versions."
          },
          "catalogVersion": { "type": "integer", "minimum": 1 },
          "policyVersion": { "type": "integer", "minimum": 1 },
          "platform": { "type": "string", "enum": ["web", "ios", "android"] },
          "cycleIndex": { "type": "integer", "minimum": 0 },
          "rails": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/HomeFeedRailPlanEntry" }
          }
        }
      },
      "HomeFeedResponse": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": { "$ref": "#/components/schemas/HomeFeed" }
        }
      }
    }
  }
}